refactor(upgrade): share forced-reinstall dir logic with installer
Address code-review findings on the forced-reinstall fix:

- Hoist target_install_dir_name / is_install_dir_for_version into
  vite_setup::install so vp-setup recognizes a `<version>+force.*` dir
  as already-installed instead of re-downloading (the installer's
  same-version check previously used `==` and missed forced-reinstall
  dirs). Co-locates the helpers with cleanup_old_versions and documents
  the contract that the forced dir name stays semver-parseable so it is
  still garbage-collected.
- CI: resolve the published `latest` via `npm view` instead of parsing
  the local build version, which can be ahead of the registry
  mid-release-cycle and would make `vp upgrade <version>` fail to
  download. Drops the `vp --version | sed | head` parse (head-under-
  pipefail footgun) and guards the temp dir with ${RUNNER_TEMP:?}.
